+++ CHANGELOG.md
@@ -67,6 +67,7 @@ These are only breaking changes for unformatted code.
 - Fix printing of uncurried application when the lhs is a function definition https://github.com/rescript-lang/rescript-compiler/pull/6084
 - Fix parsing uncurried type starting with path https://github.com/rescript-lang/rescript-compiler/pull/6089
 - Fix bigInt comparison https://github.com/rescript-lang/rescript-compiler/pull/6097
+- Fix issue with `@this` and async functions https://github.com/rescript-lang/rescript-compiler/issues/6100
 
 #### :nail_care: Polish